- Pose une question oui/non avant de supprimer une culture. - Dialogue NewCulture plus clair.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2005/Epsitec.Cresus@5056 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.Designer/Viewer.cs
@@ -488,6 +488,10 @@ namespace Epsitec.Common.Designer
 		public void DoDeleteCulture()
 		{
 			//	Supprime la culture courante.
+			string question = string.Format(Res.Strings.Dialog.DeleteCulture.Question, this.secondaryBundle.Culture.NativeName);
+			Common.Dialogs.DialogResult result = this.module.MainWindow.DialogQuestion(question);
+			if ( result != Epsitec.Common.Dialogs.DialogResult.Yes )  return;
+
 			this.module.DeleteCulture(this.secondaryBundle);
 
 			this.UpdateCultures();
